Move GuiceExtension to internal package
