+++ guice-extension/src/main/resources/META-INF/services/io.github.kaktushose.jdac.configuration.Extension
@@ -1,1 +1,1 @@
-io.github.kaktushose.jdac.guice.GuiceExtension
+io.github.kaktushose.jdac.guice.internal.GuiceExtension
